Storing route info in an ETS table
I want to change the Erlang Yaws appmod router code in src/yaws_appmod_router.erl
to use ets tables instead. This requires the implementation of a gen_server in
src/yaws_appmod_router_server.erl that create a 'protected' ets table upon
start and provides an API for adding routes by sending routes to be added to
the ets table to the server; however when reading route info, direct access
to the protected ets table can be used.