Adding multi host routing capability
It is now possible to setup different route handling
for each IP,Port pair (#sconf{}). To make this work
a separate ETS table is used for each SConf.

By storing the TableName in the #arg.opaque Map,
yaws_appmod_router can find out where to find the
appropriate route and invoke the corresponding
Handler and Middlewares.