Make sure it works with the Yaws auth/2 callback
Yaws provides two appmod callbacks: M:auth/2 and M:out/1.

The yaws_appmod_router can now work either with or without
making use of the M:auth/2 callback. To integrate yaws_appmod_router
with an existing system (that already is using the M:auth/2 callback)
you need to enable M:auth/2 callback handling. This is done
by setting config in the OpaqueMap, when setting up, e,g:

  yaws_appmod_router:enable_auth(OpaqueMap)

Since we now have enabled auth/2 callback handling, we have to handle
both the Yaws auth/2 callback and the out/1 callback in our route handler.

We can either be sloppy and just handle the out/1 callback by checking
for an #arg{} record as argument to our route handler. The auth/2 callback
will call this function with a tuple {Arg, Auth} resulting in a crash here,
which is then ignored by the yaws_appmod_router, which in its turn will
continue with the out/1 callback that will invoke this function, this time
with just the #arg{} record.

Or we define a second clause of out reoute handler that actually
implement the M:auth/2 handling, e.g by calling an existing auth/2 function
in our system or by implementing the authentication ourselves.

If we don't enable M:auth/2 handling, we can still implement authentication
in our Middleware code. See the simple example for how to implement both
these methods.